fix: remove duplicate logging (#17)
* refactor: rename module

Renamed the Go module to match the repo name.

* fix: remove duplicate logging

Errors are now wrapped and returned to the caller, rather than logged
and returned to the caller.

Propagating the error to the caller causes errors to logged by the
`main` function (if they're propagated that far). Logging them elsewhere
results in the error being logged more than once.